Moved ngrx things around
