+++ src/app/component/currencies/currencies.component.ts
@@ -1,10 +1,9 @@
 import {Component, inject} from '@angular/core';
-import {CurrencyService} from "../../service/currencyservice/currency.service";
 import {Store} from "@ngrx/store";
 import {AsyncPipe} from "@angular/common";
 import {Observable} from "rxjs";
-import {GeocoinsState} from "../../reducers/geocoins.reducers";
-import {loadGeocoins} from "../../actions/geocoins.actions";
+import {GeocoinsState} from "../../store/reducers/geocoins.reducers";
+import {loadGeocoins} from "../../store/actions/geocoins.actions";
 
 @Component({
   selector: 'app-currencies',
@@ -20,7 +19,7 @@ export class CurrenciesComponent{
   store = inject(Store)
   protected geoCoins$:Observable<GeocoinsState>
 
-  constructor(private currencyService:CurrencyService) {
+  constructor() {
     this.geoCoins$ = this.store.select('geocoins')
     this.store.dispatch(loadGeocoins())
   }
