+++ src/app/model/pet/petshop.model.ts
@@ -1,7 +1,6 @@
 import {PetModel} from "./pet.model";
 
 export interface PetShopModel {
-    "id": number,
     "pets": PetModel[],
     "ownerId": string,
     "validFor": Date
