+++ src/app/screens/workscreen/work-screen.component.ts
@@ -1,10 +1,10 @@
 import {Component, inject} from '@angular/core';
 import {LoadingComponent} from "../../component/loading/loading.component";
 import {Observable} from "rxjs";
-import {ActivityState} from "../../reducers/activity.reducers";
+import {ActivityState} from "../../store/reducers/activity.reducers";
 import {Store} from "@ngrx/store";
 import {AsyncPipe} from "@angular/common";
-import {loadWork} from "../../actions/work.actions";
+import {loadWork} from "../../store/actions/work.actions";
 import {ReactiveFormsModule} from '@angular/forms';
 import {StartWorkComponent} from "../../component/activity/start-work/start-work.component";
 import {WorkActivityComponent} from "../../component/activity/work-activity/work-activity.component";
