+++ src/app/store/actions/activity.actions.ts
@@ -1,5 +1,5 @@
 import {createAction, props} from "@ngrx/store";
-import {ActivityModel} from "../model/activity/activity.model";
+import {ActivityModel} from "../../model/activity/activity.model";
 
 export const loadActivity = createAction('[ActivityModel] Load ActivityModel')
 export const loadActivitySuccess = createAction('[ActivityModel] Load ActivityModel Success', props<{activity: ActivityModel}>())
