+++ src/app/store/actions/pets.actions.ts
@@ -1,5 +1,5 @@
 import {createAction, props} from "@ngrx/store";
-import {PetModel} from "../model/pet/pet.model";
+import {PetModel} from "../../model/pet/pet.model";
 
 export const loadPets = createAction('[Pets] Load Pets')
 export const loadPetsSuccess = createAction('[Pets] Load Pets Success', props<{pets: PetModel[]}>())
